test: Test queryObjectToQueryString.spec
